funcs to utils
